fixing i386 and x86_64 problem

+++ Repository/ulyaoth-nginx-pagespeed/build-ulyaoth-nginx-pagespeed.sh
@@ -30,8 +30,19 @@ su ulyaoth -c "unzip release-1.9.32.3-beta.zip"
 su ulyaoth -c "tar xvf 1.9.32.3.tar.gz"
 su ulyaoth -c "cp -rf ngx_pagespeed-release-1.9.32.3-beta/* /etc/nginx/modules/pagespeed/"
 su ulyaoth -c "mv psol/ /etc/nginx/modules/pagespeed/"
+
+if [ "$arch" == "x86_64" ]
+then
 su ulyaoth -c "rm -rf /etc/nginx/modules/pagespeed/psol/lib/Debug/linux/ia32"
 su ulyaoth -c "rm -rf /etc/nginx/modules/pagespeed/psol/lib/Release/linux/ia32"
+fi
+
+if [ "$arch" == "i686" ]
+then
+su ulyaoth -c "rm -rf /etc/nginx/modules/pagespeed/psol/lib/Debug/linux/x64"
+su ulyaoth -c "rm -rf /etc/nginx/modules/pagespeed/psol/lib/Release/linux/x64"
+fi
+
 su ulyaoth -c "rm -rf 1.9.32.3.tar.gz ngx_pagespeed-release-1.9.32.3-beta release-1.9.32.3-beta.zip"
 cd /etc/nginx/modules/pagespeed/
 su ulyaoth -c "tar cvf pagespeed.tar.gz scripts/ test/"
